fixed test errors, circular imports, inheritance
still getting 34 failed tests due to TypeErrors and unwrapped edges
